updated entitybulkeditservice
